agent_scratchpad/requirements.txt
requests==2.32.5
pydantic==2.12.2
loguru
python-dotenv==1.1.1
ddgs==9.5.2
